Update logo and add social media icons (#6)
This pull request contains changes primarily aimed at enhancing the user
interface and improving the accessibility of the website. The changes
include updating the website language, adding meta tags, and enhancing
the layout and design of the website.

Website Language and Metadata:
*
[`public/index.html`](diffhunk://#diff-c2cc24bc9001b11b6add48a4cd8f893d5d6c6e4d1bd254158bd14ab997f552cdL2-R25):
The language of the website has been changed from English to Danish.
Additionally, meta tags for charset, viewport, description, and keywords
have been added for better SEO. The description and keywords are
specific to the gaming guild "Semper Danica".

Website Layout and Design:
*
[`public/index.html`](diffhunk://#diff-c2cc24bc9001b11b6add48a4cd8f893d5d6c6e4d1bd254158bd14ab997f552cdL2-R25):
The logo has been updated, and a header and main section have been
added. The main section includes social media icons with links to the
respective platforms. The alt and title attributes have been added to
the image and link tags for better accessibility.
*
[`public/style.css`](diffhunk://#diff-46467272678c3da621692b13b96d47ed25128fa8027d8dbedd84684c6336fb51L1-R36):
The CSS has been updated to change the layout of the website to center
align, add a gradient background, and style the social icons. The
changes also include making the website responsive by using flexbox.

These changes aim to enhance the user experience and make the website
more accessible and SEO-friendly.